Refactor to pull logic up from SearchBox to Search
